Add proc.KillProcessGroup function.
For windows, the `killProcess` implementation was moved from the dpipe
package to the proc package.
